refactor: move task components to parent folder for shared use
