+++ src/components/task-board/task-card/TaskCardContent.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import TaskCheck from "./TaskCheck";
 import TaskName from "./TaskName";
 import { Task } from "@/types/task";
-import TaskPriority from "./TaskPriority";
-import TaskAssignee from "./TaskAssignee";
+import TaskPriority from "../TaskPriority";
+import TaskAssignee from "../TaskAssignee";
 import TaskDate from "./TaskDate";
 import { Card } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
