+++ src/components/task-board/task-sheet/TaskSheetPriority.tsx
@@ -1,7 +1,7 @@
 import { Flag } from "lucide-react";
 import React from "react";
 import LabelBadge from "./LabelBadge";
-import TaskPriority from "../task-card/TaskPriority";
+import TaskPriority from "../TaskPriority";
 import RemovePropertyButton from "./RemovePropertyButton";
 import { useTaskStore } from "@/store/useTaskStore";
